Fixed bugs in restarting; they delay convergence. Fixed C++ compiler issues. Added checks that pointer passed to magma blas/lapack functions are from device (it makes debug very slow!) Removed gotos in blaslapack. Avoided to pass uninitialized matrices to blas/lapack functions even when they are will be multiplied by zero.

+++ src/eigs/ortho.c
@@ -396,7 +396,7 @@ int Bortho_local_Sprimme(SCALAR *V, int ldV, SCALAR *R,
       primme_context ctx) {
 
    (void)ctx; 
-   struct local_matvec_ctx Bctx = {B, nLocal, ldB};
+   struct local_matvec_ctx Bctx = {B, (int)nLocal, ldB};
    return Bortho_gen_Sprimme(V, ldV, R, ldR, b1, b2, locked, ldLocked,
                              numLocked, nLocal, B ? local_matvec : NULL, &Bctx,
                              iseed, primme_get_context(NULL));
